+++ packages/admin/admin/src/router/Context.tsx
@@ -4,7 +4,12 @@ import * as React from "react";
 import { SaveAction } from "./PromptHandler";
 
 interface IContext {
-    register: (id: string, message: (location: History.Location, action: History.Action) => string | boolean, saveAction?: SaveAction) => void;
+    register: (options: {
+        id: string;
+        path: string;
+        message: (location: History.Location, action: History.Action) => string | boolean;
+        saveAction?: SaveAction;
+    }) => void;
     unregister: (id: string) => void;
 }